+++ Changelog.md
@@ -3,6 +3,8 @@
 ## V 1.4.2 - XX/07/19
 
     - Tilemap : place tile at bottom
+    
+    - Bug Fix : Propagate hide and show to child in Widgets
 
 ## V 1.4.1 - 14/07/19