+++ Changelog.md
@@ -4,9 +4,9 @@
 
     - Tilemap : place tile at bottom
     
-    - Bug Fix : Propagate hide and show to child in Widgets
-    - Bug Fix : Remove focus if widget is hide
     - Bug Fix : Remove offset when change position
+    - Bug Fix : Hide and Show doesn't spread to child in Widgets
+    - Bug Fix : Focus stays when widget is hide
 
 ## V 1.4.1 - 14/07/19