+++ Changelog.md
@@ -5,6 +5,8 @@
     - PhysicsComponent : Add callback parameters of constructor
     - Tilemap : place tile at bottom
     
+    - Optimization
+    
     - Bug Fix : OUTOFWINDOW doesn't work with CameraSystem
     - Bug Fix : Remove offset when change position
     - Bug Fix : Hide and Show doesn't spread to child in Widgets